# frontend/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hackathon API",
    "description": "Logistics matching hackathon API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://43.202.205.211",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/shippers/me": {
      "get": {
        "tags": ["Shipper"],
        "summary": "내 화주 프로필 조회",
        "operationId": "getMe",
        "parameters": [
          {
            "name": "shipperId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShipperProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Shipper"],
        "summary": "내 화주 프로필 수정",
        "operationId": "updateMe",
        "parameters": [
          {
            "name": "shipperId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipperProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShipperProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drivers/me/route-preferences": {
      "put": {
        "tags": ["Driver"],
        "summary": "다니는 구간 전체 교체 저장",
        "operationId": "updateRoutePreferences",
        "parameters": [
          {
            "name": "driverId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePreferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDriverResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/drivers/me": {
      "get": {
        "tags": ["Driver"],
        "summary": "내 기사 프로필 조회",
        "operationId": "getMe_1",
        "parameters": [
          {
            "name": "driverId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDriverResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ShipperProfileRequest": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "contactName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "businessNumber": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "address": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": ["companyName", "contactName", "phoneNumber"]
      },
      "ApiResponseShipperProfileResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ShipperProfileResponse"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ShipperProfileResponse": {
        "type": "object",
        "properties": {
          "shipperId": {
            "type": "integer",
            "format": "int64"
          },
          "companyName": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "businessNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "RegionPoint": {
        "type": "object",
        "properties": {
          "sido": {
            "type": "string"
          },
          "sigungu": {
            "type": "string"
          }
        }
      },
      "RoutePreferenceRequest": {
        "type": "object",
        "properties": {
          "origins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionPoint"
            },
            "minItems": 1
          },
          "destinations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionPoint"
            },
            "minItems": 1
          }
        },
        "required": ["destinations", "origins"]
      },
      "ApiResponseDriverResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DriverResponse"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "DriverResponse": {
        "type": "object",
        "properties": {
          "driverId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "plateNumber": {
            "type": "string"
          },
          "vehicleType": {
            "type": "string"
          },
          "capacityTon": {
            "type": "number"
          },
          "bodyType": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "totalTrips": {
            "type": "integer",
            "format": "int32"
          },
          "completionRate": {
            "type": "integer",
            "format": "int32"
          },
          "minAcceptFare": {
            "type": "integer",
            "format": "int32"
          },
          "routePreferences": {
            "$ref": "#/components/schemas/RoutePreferences"
          }
        }
      },
      "RoutePreferences": {
        "type": "object",
        "properties": {
          "origins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionPoint"
            }
          },
          "destinations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionPoint"
            }
          }
        }
      },
...
    }
  }
}